requirements.txt
gradio>=4.0
opencv-python>=4.8
scikit-learn>=1.3
PyWavelets>=1.4
numpy>=1.24
Pillow>=10.0
joblib>=1.3
matplotlib>=3.7
deepface>=0.0.84
tf-keras>=2.15.0
annoy>=1.17.3
